Add support for repeating some requests in the browser.
I've added this feature because it's a nice interim way of displaying responses nicely, before I add functionality for syntax highlighting and formatting of JSON/XML etc.  Most browsers are capable of displaying that kind of data (and developers often have tools to make browser display nicer).

However, this functionality desperately needs testing to attain some degree of confidence in its accuracy.  It is also of a size now that it warrants residing in its own module in order not to distract from the main purpose of the panel code.